feat(gui, color)
- Added the scoreboard part of Gui. Now you can freely edit the scoreboard (sidebar).
- Added remove_inventory() in Inventory Gui.
- Added gradient global function. Now you can gradient the text.

Refactor(gui, message, color)

- Refactored the gui structure.
- Refactored the message package structure.
- Refactored Hex color format from \\aRRGGBB to \aRRGGBB.

fix(inventory)

- Fixed storage issues with Inventory.

BREAKING CHANGES.

+++ src/main/kotlin/tech/ezeny/luagin/lua/api/GlobalsAPI.kt
@@ -1,6 +1,5 @@
 package tech.ezeny.luagin.lua.api
 
-import org.bukkit.Bukkit
 import party.iroiro.luajava.Lua
 import tech.ezeny.luagin.Luagin
 import tech.ezeny.luagin.lua.LuaValueFactory
@@ -134,7 +133,7 @@ object GlobalsAPI : LuaAPIProvider {
             when (key) {
                 "owtime" -> {
                     val newTime = luaState.toInteger(3)
-                    GlobalUtils.setOverworldTime(newTime.toLong())
+                    GlobalUtils.setOverworldTime(newTime)
                     return@push 0
                 }
                 "owweather" -> {
